requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
PyJWT==1.5.0
RPi.GPIO==0.6.2

Werkzeug==0.11.10
argparse==1.2.1
click==6.6

mimerender==0.6.0
itsdangerous==0.24
requests==2.12.4
six==1.10.0

xlrd==1.0.0
xlwt==1.2.0
